fix height/weight problem

+++ homework/hw1/health-record/contracts/HealthRecord.sol
@@ -6,26 +6,26 @@ contract HealthRecord {
 		string problems;
 		string medications;
 		string allergies;
-		uint weight_kg;
-		uint height_cm;
+		uint16 weight_kg;
+		uint16 height_cm;
 	}
 
 	mapping (address => Record) records;
 
-	function createRecord(string _problems, string _medications, string _allergies, uint _weight_kg, uint _height_cm) external {
+	function createRecord(string _problems, string _medications, string _allergies, uint16 _weight_kg, uint16 _height_cm) external {
 		require( (bytes(_problems).length <= 50) && (bytes(_medications).length <= 50) && (bytes(_allergies).length <= 50));
 		require( _weight_kg > 0 && _weight_kg <= 500 && _height_cm > 0 && _height_cm <= 300);
 
 		records[msg.sender] = Record(_problems, _medications, _allergies, _weight_kg, _height_cm);
 	}
 
-	function getRecord(address _address) view external returns (string _problems, string _medications, string _allergies, uint _weight_kg, uint _height_cm) {
+	function getRecord(address _address) view external returns (string _problems, string _medications, string _allergies, uint16 _weight_kg, uint16 _height_cm) {
 		Record memory rec = records[_address];
 
 		return (rec.problems, rec.medications, rec.allergies, rec.weight_kg, rec.height_cm);
 	}
 
-	function updateRecord(address _address, string _problems, string _medications, string _allergies, uint _weight_kg, uint _height_cm) external {
+	function updateRecord(address _address, string _problems, string _medications, string _allergies, uint16 _weight_kg, uint16 _height_cm) external {
 		require (msg.sender == _address);
 
 		Record storage rec = records[_address];
